float

+++ buzz_web/share/models.py
@@ -35,7 +35,7 @@ class Config(models.Model):
     number_value = models.IntegerField(verbose_name=u'值类型数值')
 
     slope_op = models.CharField(verbose_name=u'斜率类型操作', max_length=255, choices=settings.OP_CHOICES)
-    slope_value = models.DecimalField(verbose_name=u'斜率类型数值')
+    slope_value = models.FloatField(verbose_name=u'斜率类型数值')
 
     notify_roles = models.ManyToManyField(Role, verbose_name=u'告警组', null=True, blank=True)
     notify_persons = models.ManyToManyField(Person, verbose_name=u'告警人', null=True, blank=True)
